+++ lib/ui/main_app.dart
@@ -1,9 +1,12 @@
 import 'package:event_bloc/event_bloc_widgets.dart';
 import 'package:event_navigation/event_navigation.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
 import 'package:vhcsite/bloc/app_size.dart';
 import 'package:vhcsite/bloc/blog.dart';
 import 'package:vhcsite/bloc/navigation/navigation_bloc.dart';
+import 'package:vhcsite/model/environment.dart';
 import 'package:vhcsite/repository/url_repository.dart';
 import 'package:vhcsite/ui/screens/main_screen.dart';
 import 'package:vhcsite/ui/theme/theme.dart';
@@ -13,18 +16,21 @@ class MainApp extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return RepositoryProvider(
-      create: (_) => UrlRepository(),
-      child: MultiBlocProvider(
-        blocBuilders: [
-          BlocBuilder<AppSizeBloc>((reader, parentChannel) =>
-              AppSizeBloc(parentChannel: parentChannel)),
-          BlocBuilder<MainNavigationBloc<String>>(
-              (reader, parentChannel) => generateNavigationBloc(parentChannel)),
-          BlocBuilder<BlogBloc>((reader, parentChannel) =>
-              BlogBloc(parentChannel: parentChannel)),
-        ],
-        child: _InnerApp(),
+    return Provider(
+      create: (context) => const VHCEnvironment(overrideTextScale: kIsWeb),
+      child: RepositoryProvider(
+        create: (_) => UrlRepository(),
+        child: MultiBlocProvider(
+          blocBuilders: [
+            BlocBuilder<AppSizeBloc>((reader, parentChannel) =>
+                AppSizeBloc(parentChannel: parentChannel)),
+            BlocBuilder<MainNavigationBloc<String>>((reader, parentChannel) =>
+                generateNavigationBloc(parentChannel)),
+            BlocBuilder<BlogBloc>((reader, parentChannel) =>
+                BlogBloc(parentChannel: parentChannel)),
+          ],
+          child: _InnerApp(),
+        ),
       ),
     );
   }