+++ lib/ui/watcher/watcher.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+
+class WatcherLayer extends StatelessWidget {
+  const WatcherLayer({super.key, required this.child});
+  final Widget child;
+
+  @override
+  Widget build(BuildContext context) {
+    return child;
+  }
+}